+++ sections/catalog.php
@@ -3,23 +3,21 @@
 <head>
     <?php
     $sectionTitle = "GeekStore - Catálogo";
-    require_once('header.php');
+    require_once('parts/header.php');
     ?>
-    <link rel="stylesheet" href="../style/catalog.css">
-    <style>
-        a[href='catalog.php']{
-            color: white;
-            background-color: black;
-        }
-    </style>
+    <link rel="stylesheet" href="../style/sections/catalog.css">
 </head>
 <body>
-<?php require_once('navbar.php'); ?>
+
+<?php require_once('parts/navbar.php'); ?>
+
 <div class="box-container">
     <input id="search" class="search-box" type="text" name="search" onkeyup="search()" placeholder="Buscar..">
 </div>
 <div id="catalog"></div>
-<?php require_once('scripts.php'); ?>
+
+<?php require_once('parts/scripts.php'); ?>
+
 <script src="../javascript/sections/catalog.js"></script>
 </body>
 </html>
