+++ sections/login.php
@@ -3,31 +3,28 @@
 <head>
     <?php
     $sectionTitle = "GeekStore - Iniciar Sesión";
-    require_once('header.php');
+    require_once('parts/header.php');
     ?>
     <link rel="stylesheet" href="../style/forms.css">
-    <style>
-        a[href='login.php']{
-            background-color: black;
-            color : white;
-        }
-    </style>
+    <link rel="stylesheet" href="../style/sections/login.css">
 </head>
 <body>
-<?php require_once('navbar.php'); ?>
+
+<?php require_once('parts/navbar.php'); ?>
+
 <div class="container">
   <div class="form">
     <h1>Iniciar Sesión</h1>
     <form action="" name="form" method="POST">
-        <input type="text" name="username" id="username" placeholder="Nombre de usuario"><br>
-        <input type="password" name="password" id="password" placeholder="Contrase&ntilde;a"><br>
-        <input type="button" value="Iniciar" class="button" id="btnLogin">
+        <input type="text" name="username" id="username" class="form-input" placeholder="Nombre de usuario"><br>
+        <input type="password" name="password" id="password" class="form-input" placeholder="Contrase&ntilde;a"><br>
+        <input type="button" value="Iniciar" class="geek-btn" id="btnLogin">
     </form>
   </div>
 </div>
-<?php
-require_once('scripts.php');
-?>
+
+<?php require_once('parts/scripts.php'); ?>
+
 <script src="../javascript/sections/login.js"></script>
 </body>
 </html>
