+++ sections/parts/scripts.php
@@ -1,0 +1,2 @@
+<script src="../../javascript/notifier.js"></script>
+<script src="../../javascript/navbar.js"></script>
