--- sections/scripts.php
@@ -1,2 +1,0 @@
-<script src="../javascript/notifier.js"></script>
-<script src="../javascript/navbar.js"></script>
