+++ style/forms.css
@@ -1,15 +1,6 @@
-@font-face {
-    font-family: game;
-    src: url('../resources/fonts/game/PressStart2P-Regular.ttf');
-}
-body {
-    font-family: game, serif;
-    background: black;
-}
 input {
     font-family: game, serif;
-    border: none;
-    border-bottom: none;
+    border-style: none;
     margin: 10px;
     outline: none;
 }
@@ -38,11 +29,10 @@ input:focus {
 .form form {
     overflow: hidden;
     height: auto;
-    overflow: hidden;
     margin-left: auto;
     margin-right: auto;
 }
-.form form input[type=text], .form form input[type=password], .form form input[type=tel] {
+.form .form-input{
     width: 100%;
     position: relative;
     font-size: 12px;
@@ -52,7 +42,7 @@ input:focus {
     border: none;
     color: #000;
     outline: none;
-    margin: 0 auto;
+    margin: 1% auto;
 }
 .form .button {
     padding: 14px;
@@ -66,6 +56,7 @@ input:focus {
     float: right;
     margin-right: inherit;
 }
+
 ::-webkit-input-placeholder {
     /* WebKit, Blink, Edge */
     color: midnightblue;
@@ -86,9 +77,8 @@ input:focus {
     color: midnightblue;
 }
 
-.form form input[type=text], .form form input[type=password], .form form input[type=tel],
-::-webkit-input-placeholder, .container h1, .form .button  {
+.form .form-input,
+::-webkit-input-placeholder,.container h1{
   transform:scale(1,1.2); /* W3C */
   -webkit-transform:scale(1,1.2); /* Safari and Chrome */
-
 }
