+++ style/header.css
@@ -1,10 +1,3 @@
-@font-face {
-    font-family: game;
-    src: url('../resources/fonts/game/PressStart2P-Regular.ttf');
-}
-body{
-    font-family: game, serif;
-}
 #header{
     background-color: midnightblue;
 }
