+++ style/home.css
@@ -1,8 +1,14 @@
+@font-face {
+    font-family: game;
+    src: url('../resources/fonts/game/PressStart2P-Regular.ttf');
+}
 body{
+    font-family: game, serif;
     margin: 0;
     padding: 0;
     width: 100%;
     height: 100%;
+    background-color: black;
 }
 
 #principal{
@@ -22,4 +28,38 @@ p{
 }
 .right{
     float: right;
+}
+
+.geek-btn {
+    margin: 5px;
+    padding-top: 2%;
+    padding-bottom: 2%;
+    border: none;
+    color: whitesmoke;
+    background-color: midnightblue;
+    font-size: 1.3em;
+    -webkit-box-shadow: 5px 5px 5px black;
+    -moz-box-shadow: 5px 5px 5px black;
+    box-shadow: 5px 5px 5px black;
+    -webkit-border-radius:5%;
+    -moz-border-radius:5%;
+    border-radius:5%;
+    float: right;
+    letter-spacing: 1px;
+    font-family: game, sans-serif;
+    transform:scale(1,1.2); /* W3C */
+    -webkit-transform:scale(1,1.2); /* Safari and Chrome */
+}
+
+.geek-btn:hover{
+    background-color: dodgerblue;
+    cursor: pointer;
+}
+
+.geek-btn:focus{
+    background-color: midnightblue;
+    font-size: 1.3em;
+    -webkit-box-shadow: 0 0 0 0 #FFFFFF;
+    -moz-box-shadow: 0 0 0 0 #FFFFFF;
+    box-shadow: 0 0 0 0 #FFFFFF;
 }
