+++ style/sections/catalog.css
@@ -1,12 +1,15 @@
 @font-face {
     font-family: game;
-    src: url('../resources/fonts/game/PressStart2P-Regular.ttf');
+    src: url('../../resources/fonts/game/PressStart2P-Regular.ttf');
 }
 body {
     font-family: game, serif;
     background: black;
 }
-
+a[href='catalog.php']{
+    color: white;
+    background-color: black;
+}
 input {
     font-family: game, serif;
 }
@@ -14,15 +17,16 @@ input {
 #catalog, .box-container {
     width: 100%;
     margin-top: 2%;
-    margin-left: 5%;
-    margin-right: 5%;
+    padding-left: 5%;
+    padding-right: 5%;
+    border: 0;
+    margin-left: 0;
+    margin-right: 0;
+
 }
 
 .box-container input[type=text] {
-    background-color: white;
-    background-image: url('../resources/images/icons/searchicon.png');
-    background-position: 10px 10px;
-    background-repeat: no-repeat;
+    background: white url('../../resources/images/icons/searchicon.png') no-repeat 10px 10px;
     -webkit-transition: width 0.3s ease-in-out;
     transition: width 0.3s ease-in-out;
     width: 20%;
@@ -38,12 +42,12 @@ input {
 }
 
 .shadow-depth-light {
-   box-shadow: 0 1px 3px rgba(255,255,255, 0.12), 0 1px 2px rgba(255,255,255,0.24);
+    box-shadow: 0 1px 3px rgba(255,255,255, 0.12), 0 1px 2px rgba(255,255,255,0.24);
 }
 
 .border-tlr-radius {
     border-top-left-radius: 2px;
-	  border-top-right-radius: 2px;
+    border-top-right-radius: 2px;
 }
 
 .card {
@@ -92,8 +96,8 @@ input {
 
 .card-image img {
     width: 100%;
-		max-width: 100%;
-		display: block;
+    max-width: 100%;
+    display: block;
 }
 
 .card-content {
@@ -102,7 +106,7 @@ input {
 
 .action-button {
     display: inline-block;
-    border-top: 1.5px solid midnightblue;
+    border-top: 1px solid midnightblue;
     border-bottom: none;
     border-right: none;
     border-left: none;
@@ -115,6 +119,7 @@ input {
 .action-button:hover {
     background: yellow;
     color: white;
+    cursor: pointer;
 }
 
 .button-image {
