+++ style/sections/contact.css
@@ -2,12 +2,12 @@ html, body { height: 100%; margin: 0; padding: 0; }
 
 @font-face {
     font-family: game;
-    src: url('../resources/fonts/game/PressStart2P-Regular.ttf');
+    src: url('../../resources/fonts/game/PressStart2P-Regular.ttf');
 }
 
 body{
     font-family: game, serif;
-    background: lightgrey url('../resources/images/background/pacman2.jpg') center;
+    background: lightgrey url('../../resources/images/background/pacman2.jpg') center;
 }
 
 #map {
