+++ style/sections/login.css
@@ -1,0 +1,7 @@
+a[href='login.php']{
+    background-color: black;
+    color : white;
+}
+body{
+    background-color: black;
+}
