+++ style/sections/manage_product.css
@@ -1,0 +1,68 @@
+a[href='manage_product.php']{
+    color: white;
+    background-color: black;
+}
+body{
+    background-color: black;
+}
+#products tr:hover{
+    cursor: pointer;
+}
+#products tr th{
+    background-color: darkgray;
+    padding: 1%;
+}
+#products table{
+    margin-top: 5%;
+}
+#products table button{
+    padding: 10px;
+    border: 0;
+    background-color: red;
+    color: white;
+    margin: 2px;
+}
+#products table button:hover{
+    cursor: pointer;
+    background-color: darkred;
+}
+#products tr td{
+    text-align: center;
+}
+#products tr:nth-child(odd) {
+    background-color: white;
+}
+#products tr:nth-child(even) {
+    background-color: lightgrey;
+}
+#products{
+    margin-left: 20%;
+    margin-right: 20%;
+    width: 60%;
+}
+#modal_window_body{
+    margin: 1%;
+}
+#modal_window_body form input{
+    border-bottom: 2px solid blue;
+}
+#modal_window_body label{
+    font-size: 0.6em;
+}
+#modal_window_body input{
+    font-size: 0.5em;
+}
+#modal_window_body textarea{
+    border: none;
+    width: 70%;
+    height: 80px;
+    font-family: game, sans-serif;
+    font-size: 0.4em;
+    border-bottom: 2px solid blue;
+}
+#description-wrapper label{
+    display: block;
+    vertical-align: middle;
+    margin-top: 1%;
+    margin-bottom: 1%;
+}
