+++ style/sections/manage_user.css
@@ -1,0 +1,54 @@
+body{
+    background-color: black;
+}
+a[href='manage_user.php']{
+    color: white;
+    background-color: black;
+}
+#users tr:hover{
+    cursor: pointer;
+}
+#users tr th{
+    background-color: darkgray;
+    padding: 1%;
+}
+#users table{
+    margin-top: 5%;
+}
+#users table button{
+    padding: 10px;
+    border: 0;
+    background-color: red;
+    color: white;
+    margin: 2px;
+}
+#users table button:hover{
+    cursor: pointer;
+    background-color: darkred;
+}
+#users tr td{
+    text-align: center;
+}
+#users tr:nth-child(odd) {
+    background-color: white;
+}
+#users tr:nth-child(even) {
+    background-color: lightgrey;
+}
+#users{
+    margin-left: 20%;
+    margin-right: 20%;
+    width: 60%;
+}
+#modal_window_body form input{
+    border-bottom: 2px solid blue;
+}
+#modal_window_body select{
+    padding: 0.8%;
+}
+#modal_window_body label{
+    font-size: 0.6em;
+}
+#modal_window_body input{
+    font-size: 0.5em;
+}
