+++ style/sections/profile.css
@@ -1,3 +1,7 @@
+a[href='profile.php']{
+    color: white;
+    background-color: black;
+}
 body{
     background: black;
 }
