+++ style/sections/register.css
@@ -1,0 +1,4 @@
+a[href='register.php']{
+    background-color: black;
+    color : white;
+}
